# auth-service/main-app/docs/swagger.yaml
basePath: /api/
definitions:
  models.Login:
    properties:
      email:
        type: string
      password:
        type: string
      role:
        enum:
        - Issuer
        - Verifier
        - User
        type: string
    required:
    - email
    - password
    type: object
  models.OTP:
    properties:
      email:
        type: string
      otp:
        type: string
    required:
    - email
    - otp
    type: object
  models.Register:
    properties:
      email:
        type: string
      password:
        type: string
      role:
        enum:
        - Issuer
        - Verifier
        - User
        type: string
    required:
    - email
    - password
    type: object
  responses.ErrorResponse_doc:
    properties:
      message:
        type: string
    type: object
  responses.UserResponse_doc:
    properties:
      data:
        additionalProperties: true
        type: object
      message:
        type: string
    type: object
info:
  contact: {}
  description: This is an auth api for an application.
  title: Auth API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Allows users to login into their account.
      parameters:
      - description: User's email and password
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.Login'
      produces:
      - application/json
      responses:
        "200":
          description: Successful response
          schema:
            $ref: '#/definitions/responses.UserResponse_doc'
        "400":
          description: Please provide with sufficient credentials
          schema:
            $ref: '#/definitions/responses.ErrorResponse_doc'
        "401":
          description: Invalid Credentials
          schema:
            $ref: '#/definitions/responses.ErrorResponse_doc'
        "404":
          description: User is not registered
          schema:
            $ref: '#/definitions/responses.ErrorResponse_doc'
        "422":
          description: Email already registered, please verify your email address
          schema:
            $ref: '#/definitions/responses.ErrorResponse_doc'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse_doc'
      summary: Login route
      tags:
      - user
  /auth/otp:
    post:
      consumes:
      - application/json
      description: Allows users to validate OTP and complete the registration process.
      parameters:
      - description: User's email address and otp
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.OTP'
      produces:
      - application/json
      responses:
        "200":
          description: Successful response, User already verified. Please login.
          schema:
            $ref: '#/definitions/responses.UserResponse_doc'
        "400":
          description: Invalid JSON data, Invalid Email
          schema:
            $ref: '#/definitions/responses.ErrorResponse_doc'
        "401":
          description: Invalid OTP
          schema:
            $ref: '#/definitions/responses.ErrorResponse_doc'
        "404":
          description: User does not exist. Please register to generate OTP.
          schema:
            $ref: '#/definitions/responses.ErrorResponse_doc'
        "422":
          description: Please provide with sufficient credentials
          schema:
            $ref: '#/definitions/responses.ErrorResponse_doc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponse_doc'
      summary: Validation route
      tags:
      - user
  /auth/register:
    post:
      consumes:
      - application/json
      description: Allows users to create a new account.
      parameters:
      - description: User name, email, password
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.Register'
      produces:
      - application/json
      responses:
        "201":
          description: Successful response
          schema:
            $ref: '#/definitions/responses.UserResponse_doc'
        "400":
          description: Invalid JSON data, Invalid Email
          schema:
            $ref: '#/definitions/responses.ErrorResponse_doc'
        "401":
          description: Invalid Credentials
          schema:
            $ref: '#/definitions/responses.ErrorResponse_doc'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/responses.ErrorResponse_doc'
        "422":
          description: Please provide with sufficient credentials
          schema:
            $ref: '#/definitions/responses.ErrorResponse_doc'
        "500":
          description: Internal Server Error, Error in inserting the document
          schema:
            $ref: '#/definitions/responses.ErrorResponse_doc'
      summary: Register route
      tags:
      - user
swagger: "2.0"